issue: 서치 - 테이블뷰 버튼 작업
