+++ SeoulMate/SeoulMate/Source/Modules/Search/SearchResultTableViewCell.swift
@@ -9,10 +9,6 @@ import UIKit
 
 class SearchResultTableViewCell: UITableViewCell {
 
-
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var subTitleLabel: UILabel!
-    
     override func awakeFromNib() {
         super.awakeFromNib()
     }
